Rename zipline-tools to zipline-profiler
Also start publishing the artifact for the profiler.
