+++ settings.gradle.kts
@@ -6,9 +6,9 @@ include(":zipline-gradle-plugin")
 include(":zipline-kotlin-plugin")
 include(":zipline-kotlin-plugin-hosted")
 include(":zipline-kotlin-plugin:tests")
-include(":zipline-tools")
 include(":zipline-bytecode")
 include(":zipline-loader")
+include(":zipline-profiler")
 
 include(":samples:emoji-search")
 include(":samples:emoji-search:presenters")
