--- zipline-tools/build.gradle.kts
@@ -1,11 +1,0 @@
-plugins {
-  kotlin("jvm")
-}
-
-dependencies {
-  api(project(":zipline"))
-  api(Dependencies.okio)
-
-  testImplementation(Dependencies.truth)
-  testImplementation(Dependencies.junit)
-}
